Converted from double to float

+++ src/optimizationSolver/cost.cpp
@@ -6,17 +6,17 @@
 #include "cost.h"
 
 
-double getCost(std::vector<Building> buildings, bool volumeToggle, bool sunToggle){
-    double cost = 0;
+float getCost(std::vector<Building> buildings, bool volumeToggle, bool sunToggle){
+    float cost = 0;
     if (volumeToggle) {
-        std::vector<double> buildingAreas;
+        std::vector<float> buildingAreas;
         buildingAreas.reserve(buildings.size());
         std::transform(buildings.begin(), buildings.end(), buildingAreas.begin(), simpleArea);
-        double totalArea = std::accumulate(buildingAreas.begin(), buildingAreas.end(), 0.0);
+        float totalArea = std::accumulate(buildingAreas.begin(), buildingAreas.end(), 0.0);
         cost += totalArea;
     }
     if (sunToggle){
-        double sunCost = 0;
+        float sunCost = 0;
         cost += sunCost;
     }
     return cost;
