+++ src/optimizationSolver/cost.h
@@ -8,6 +8,6 @@
 #include "geometry.h"
 
 
-double getCost(std::vector<Building> buildings, bool volumeToggle, bool sunToggle);
+float getCost(std::vector<Building> buildings, bool volumeToggle, bool sunToggle);
 
 #endif //OPTIMIZE_COST_H
